course22/requirements.txt
Django>=2.2
django-debug-toolbar>=1.9.1
pytz
sqlparse
wheel
django-queryinspect>=1.0.0

djangorestframework>=3.10.3
raven
